+++ test/features/dive_import/data/services/uddf_entity_importer_test.mocks.dart
@@ -2058,6 +2058,16 @@ class MockDiveRepository extends _i1.Mock implements _i13.DiveRepository {
           as _i17.Future<List<_i35.DiveDataSource>>);
 
   @override
+  _i17.Future<Map<String, String>> getSourceUuidByDiveId({String? diverId}) =>
+      (super.noSuchMethod(
+            Invocation.method(#getSourceUuidByDiveId, [], {#diverId: diverId}),
+            returnValue: _i17.Future<Map<String, String>>.value(
+              <String, String>{},
+            ),
+          )
+          as _i17.Future<Map<String, String>>);
+
+  @override
   _i17.Future<bool> hasMultipleDataSources(String? diveId) =>
       (super.noSuchMethod(
             Invocation.method(#hasMultipleDataSources, [diveId]),