+++ test/features/dive_import/presentation/providers/dive_import_notifier_test.mocks.dart
@@ -588,6 +588,16 @@ class MockDiveRepository extends _i1.Mock implements _i3.DiveRepository {
           as _i5.Future<List<_i12.DiveDataSource>>);
 
   @override
+  _i5.Future<Map<String, String>> getSourceUuidByDiveId({String? diverId}) =>
+      (super.noSuchMethod(
+            Invocation.method(#getSourceUuidByDiveId, [], {#diverId: diverId}),
+            returnValue: _i5.Future<Map<String, String>>.value(
+              <String, String>{},
+            ),
+          )
+          as _i5.Future<Map<String, String>>);
+
+  @override
   _i5.Future<bool> hasMultipleDataSources(String? diveId) =>
       (super.noSuchMethod(
             Invocation.method(#hasMultipleDataSources, [diveId]),